+++ src/main/java/katecam/hyuswim/user/controller/UserController.java
@@ -2,15 +2,14 @@ package katecam.hyuswim.user.controller;
 
 import katecam.hyuswim.common.ApiResponse;
 import katecam.hyuswim.user.User;
-import katecam.hyuswim.user.dto.JwtTokenRequest;
-import katecam.hyuswim.user.dto.JwtTokenResponse;
+import katecam.hyuswim.jwt.dto.JwtTokenRequest;
+import katecam.hyuswim.jwt.dto.JwtTokenResponse;
 import katecam.hyuswim.user.dto.LoginRequest;
 import katecam.hyuswim.user.dto.SignupRequest;
 import katecam.hyuswim.user.exception.UserNotFoundException;
-import katecam.hyuswim.user.jwt.JwtUtil;
+import katecam.hyuswim.jwt.JwtUtil;
 import katecam.hyuswim.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
